# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comparador de Normativas — API REST & SSE",
...
    "version": "4.0.0"
  },
  "paths": {
...
    "/api/index/build": {
      "post": {
        "tags": [
          "Índice Semántico"
        ],
        "summary": "Build Index",
        "description": "Construye el índice semántico FAISS con o sin sub-chunking parent-child.",
        "operationId": "build_index_api_index_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexBuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/index/search": {
      "post": {
        "tags": [
          "Índice Semántico"
        ],
        "summary": "Search Index",
        "description": "Ejecuta una búsqueda semántica de prueba.",
        "operationId": "search_index_api_index_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndexBuildRequest": {
        "properties": {
          "use_chunking": {
            "type": "boolean",
            "title": "Use Chunking",
            "default": true
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "default": 512
          },
          "solape": {
            "type": "integer",
            "title": "Solape",
            "default": 50
          }
        },
        "type": "object",
        "title": "IndexBuildRequest"
      },
      "IndexBuildResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_vectores": {
            "type": "integer",
            "title": "Total Vectores"
          },
          "usar_chunking": {
            "type": "boolean",
            "title": "Usar Chunking"
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds"
          }
        },
        "type": "object",
        "required": [
          "status",
          "total_vectores",
          "usar_chunking",
          "elapsed_seconds"
        ],
        "title": "IndexBuildResponse"
      },
...
      "SearchQueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 5
          },
          "min_score": {
            "type": "number",
            "title": "Min Score",
            "default": 0.3
          },
          "use_reranker": {
            "type": "boolean",
            "title": "Use Reranker",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchQueryRequest"
      },
      "SearchQueryResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results"
        ],
        "title": "SearchQueryResponse"
      },
      "SearchResultItem": {
        "properties": {
          "element_id": {
            "type": "string",
            "title": "Element Id"
          },
          "numero": {
            "type": "string",
            "title": "Numero"
          },
          "encabezado": {
            "type": "string",
            "title": "Encabezado"
          },
          "contenido": {
            "type": "string",
            "title": "Contenido"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          }
        },
        "type": "object",
        "required": [
          "element_id",
          "numero",
          "encabezado",
          "contenido",
          "score",
          "doc_id"
        ],
        "title": "SearchResultItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}